+++ src/resourcetypes/controllers.js
@@ -1,0 +1,10 @@
+function ManageCtrl() {
+}
+
+function ListCtrl() {
+}
+
+
+angular.module('md.resourcetypes')
+  .controller('ManageCtrl', ManageCtrl)
+  .controller('ListCtrl', ListCtrl);